requirements.txt
google-generativeai>=0.2.0
pdfplumber>=0.6.0
pydantic>=1.10
pillow>=9.0.0
pdf2image>=1.16.0
boto3>=1.26.0
pypdf
